feat: mempool validates `fee` and `value` fields for stakes and pledges (#576)
* feat: validation for fee & value for commitment txs
* feat: pledge count is included in the commitment tx
* feat: update constructor of commitment txs to take the fee from the config
* test: invalid stake txs get rejected
* feat: explicit error for invalid commitments

+++ crates/types/src/storage_pricing.rs
@@ -223,7 +223,7 @@ impl Amount<Irys> {
     #[tracing::instrument(err)]
     pub fn apply_pledge_decay(
         self,
-        current_pledge_count: usize,
+        current_pledge_count: u64,
         decay_rate: Amount<Percentage>,
     ) -> Result<Self> {
         // The formula is: pledge_value = pledge_base_fee / ((count + 1) ^ decay_rate)
